@@ -1,6 +1,6 @@
 # This file is copied to spec/ when you run 'rails generate rspec:install'
 ENV["RAILS_ENV"] ||= 'test'
-require File.expand_path("../../config/environment", __FILE__)
+require_relative "../config/environment"
 require 'rspec/rails'
 require 'clearance/rspec'
 require 'turnip'
