@@ -21,12 +21,12 @@ Capybara.register_driver :chrome do |app|
 end
 
 Capybara.register_driver :headless_chrome do |app|
-  caps = Selenium::WebDriver::Remote::Capabilities.chrome(loggingPrefs: { browser: 'ALL' })
-  opts = Selenium::WebDriver::Chrome::Options.new
+  opts = Selenium::WebDriver::Options.chrome
+  opts.add_option("goog:loggingPrefs", { browser: "ALL" })
 
   chrome_args = %w[--headless --window-size=1920,1080 --no-sandbox --disable-dev-shm-usage]
   chrome_args.each { |arg| opts.add_argument(arg) }
-  Capybara::Selenium::Driver.new(app, browser: :chrome, options: opts, desired_capabilities: caps)
+  Capybara::Selenium::Driver.new(app, browser: :chrome, options: opts)
 end
 
 Capybara.configure do |config|
